+++ lib/qig.rb
@@ -52,13 +52,13 @@ module Qig
       end
     end
 
-    def collection_qig(subjects, *path) # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
+    def collection_qig(subjects, *path) # rubocop:disable Metrics/AbcSize, Metrics/MethodLength, Metrics/CyclomaticComplexity
       head, *rest = path
       case head
       in nil
         subjects
       in []
-        collection_qig(subjects.map(&method(:values)).flatten(1), *rest)
+        collection_qig(subjects.map(&method(:values)).flat_map(&:itself), *rest)
       in [[]]
         unit_qig(subjects, *rest)
       in ['', key]