+++ lib/qig/qiggable.rb
@@ -2,10 +2,10 @@
 
 module Qig
   # mix-in to add a `.qig` method to any class.
-  # NOTE: Target class should support `#[]`, `#map`, and either `#flatten` or `#values` for full qig functionality
+  # NOTE: Target class should support `#[]`, `#map`, and either `#flat_map` or `#values` for full qig functionality
   # - `#[]`: used for stepping in unit context
   # - `#map`: used for stepping in collection context
-  # - `#flatten` or `#values`: used for unboxing in collection context
+  # - `#flat_map` or `#values`: used for unboxing in collection context
   module Qiggable
     # @param path [Array<String, Symbol, Array, Object>] retrieval path to apply to `subject`
     #